SEO_Work-main/requirements.txt
streamlit
beautifulsoup4
requests
langchain
langchain_core
langchain-community
langchain-openai
langgraph
openai
tabulate
openpyxl
lxml
python-dotenv
urllib3
crawl4ai
html2text
nest_asyncio
playwright
datasketch
reportlab


accelerate==1.8.0
fastapi==0.115.12
fastparquet==2024.11.0
faiss-cpu==1.12.0
huggingface-hub==0.33.1
keybert==0.9.0
numpy==1.26.4
pandas==2.2.3
pydantic==2.11.5
regex==2024.11.6
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==2.7.0
sentencepiece==0.2.0 
tokenizers==0.19.1
torch==2.2.2
tqdm==4.67.1
transformers==4.40.1
yake==0.7.3
uvicorn==0.32.1
XlsxWriter>=3.1.0
